Make FileWriter in ComparisonReportWriter exchangeable
